major changes made to code, breaking out into new components
