+++ src/components/MovieCard/MovieCard.scss
@@ -1,0 +1,19 @@
+.movie-card {
+  border: red 3px solid;
+  height: 575px;
+  width: 350px;
+  margin: 15px;
+}
+
+.movie-card-image {
+  height: 400px;
+  width: 350px;
+}
+
+.rating-dropdown {
+  display: inline;
+}
+
+.rating-label {
+  display: inline;
+}
